requirements.txt
asgiref==3.2.10
Django==3.1.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.5
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
